refactor: hoist server calls into top-level app/src/api/
manifestStream → api/manifest, serverConfig → api/config,
url → api/urls, views/panes/commitFetch → api/commit. No behavior
change — only file paths.
