+++ app/tests/api/config.test.ts
@@ -1,9 +1,5 @@
 import { describe, it, expect, beforeEach, vi } from 'vitest';
-import {
-  fetchServerConfig,
-  getServerConfig,
-  _resetServerConfigForTests,
-} from '@/utils/serverConfig.js';
+import { fetchServerConfig, getServerConfig, _resetServerConfigForTests } from '@/api/config.js';
 
 describe('fetchServerConfig', () => {
   beforeEach(() => {
